+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Editing:
     * Added support for precompositions
+    * Ability to copy/paste keyframe values (through menus)
 * UI:
     * The timeline now allows editing layer frame ranges
     * When selecting groups, its inner objects will show in the timeline